+++ src/lib.rs
@@ -134,7 +134,7 @@
 //!
 //! ```rust,no_run
 //! use embedded_graphics::{
-//!     mono_font::{Font6x8, MonoTextStyle},
+//!     mono_font::{ascii::Font6x9, MonoTextStyle},
 //!     pixelcolor::BinaryColor,
 //!     prelude::*,
 //!     primitives::{Circle, Rectangle, Triangle},
@@ -152,7 +152,7 @@
 //!     let thin_stroke = PrimitiveStyle::with_stroke(BinaryColor::On, 1);
 //!     let thick_stroke = PrimitiveStyle::with_stroke(BinaryColor::On, 3);
 //!     let fill = PrimitiveStyle::with_fill(BinaryColor::On);
-//!     let text_style = MonoTextStyle::new(Font6x8, BinaryColor::On);
+//!     let text_style = MonoTextStyle::new(Font6x9, BinaryColor::On);
 //!
 //!     let yoffset = 10;
 //!