+++ src/mono_font/mod.rs
@@ -8,7 +8,7 @@
 //!
 //! # Examples
 //!
-//! The examples below use the [`Font6x8`] font, however any of the [built-in fonts]
+//! The examples below use the `Font6x8` font, however any of the [built-in fonts]
 //! in this module or custom fonts can be substituted.
 //!
 //! ## Print styled "Hello Rust!"
@@ -19,7 +19,7 @@
 //!
 //! ```rust
 //! use embedded_graphics::{
-//!     mono_font::{Font6x8, MonoTextStyle, MonoTextStyleBuilder},
+//!     mono_font::{ascii::Font6x9, MonoTextStyle, MonoTextStyleBuilder},
 //!     pixelcolor::Rgb565,
 //!     prelude::*,
 //!     text::Text,
@@ -30,7 +30,7 @@
 //!
 //! // Create a new text style
 //! let style = MonoTextStyleBuilder::new()
-//!     .font(Font6x8)
+//!     .font(Font6x9)
 //!     .text_color(Rgb565::YELLOW)
 //!     .background_color(Rgb565::BLUE)
 //!     .build();
@@ -46,7 +46,7 @@
 //!
 //! ```rust
 //! use embedded_graphics::{
-//!     mono_font::{Font6x8, MonoTextStyle},
+//!     mono_font::{ascii::Font6x9, MonoTextStyle},
 //!     pixelcolor::BinaryColor,
 //!     prelude::*,
 //!     text::Text,
@@ -56,7 +56,7 @@
 //! # display.set_allow_out_of_bounds_drawing(true);
 //!
 //! Text::new("Hello Rust!", Point::zero())
-//!     .into_styled(MonoTextStyle::new(Font6x8, BinaryColor::On))
+//!     .into_styled(MonoTextStyle::new(Font6x9, BinaryColor::On))
 //!     .translate(Point::new(20, 30))
 //!     .draw(&mut display)?;
 //!
@@ -65,7 +65,7 @@
 //! # let mut display: MockDisplay<BinaryColor> = MockDisplay::default();
 //! # display.set_allow_out_of_bounds_drawing(true);
 //! Text::new("Hello Rust!", Point::new(20, 30))
-//!     .into_styled(MonoTextStyle::new(Font6x8, BinaryColor::On))
+//!     .into_styled(MonoTextStyle::new(Font6x9, BinaryColor::On))
 //!     .draw(&mut display)?;
 //! # Ok::<(), core::convert::Infallible>(())
 //! ```
@@ -80,7 +80,7 @@
 //! use arrayvec::ArrayString;
 //! use core::fmt::Write;
 //! use embedded_graphics::{
-//!     mono_font::{Font6x8, MonoTextStyleBuilder},
+//!     mono_font::{ascii::Font6x9, MonoTextStyleBuilder},
 //!     pixelcolor::Rgb565,
 //!     prelude::*,
 //!     text::Text,
@@ -100,7 +100,7 @@
 //! Text::new(&buf, Point::zero())
 //!     .into_styled(
 //!         MonoTextStyleBuilder::new()
-//!             .font(Font6x8)
+//!             .font(Font6x9)
 //!             .text_color(Rgb565::YELLOW)
 //!             .background_color(Rgb565::BLUE)
 //!             .build(),
@@ -111,41 +111,30 @@
 //!
 //! # Built-in fonts
 //!
+//! TODO: Replace with generated fonts table
+//!
 //! | Type | Screenshot |
 //! |------|------------|
-//! | [`Font6x8`] | ![6x8 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font6x8.png) |
-//! | [`Font6x12`] | ![6x12 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font6x12.png) |
-//! | [`Font8x16`] | ![8x16 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font8x16.png) |
-//! | [`Font12x16`] | ![12x16 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font12x16.png) |
-//! | [`Font24x32`] | The 24x32 font is a pixel doubled version of the 12x16 font. |
+//! | `Font6x8` | ![6x8 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font6x8.png) |
+//! | `Font6x12` | ![6x12 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font6x12.png) |
+//! | `Font8x16` | ![8x16 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font8x16.png) |
+//! | `Font12x16` | ![12x16 font spritemap screenshot](https://raw.githubusercontent.com/embedded-graphics/embedded-graphics/master/data/font12x16.png) |
+//! | `Font24x32` | The 24x32 font is a pixel doubled version of the 12x16 font. |
 //!
 //! [built-in fonts]: #built-in-fonts
-//! [`Font6x8`]: struct.Font6x8.html
-//! [`Font6x12`]: struct.Font6x12.html
-//! [`Font8x16`]: struct.Font8x16.html
-//! [`Font12x16`]: struct.Font12x16.html
-//! [`Font24x32`]: struct.Font24x32.html
 //! [`Text`]: ../text/struct.Text.html
 //! [`Styled`]: ../style/struct.Styled.html
 //! [`MonoTextStyle`]: struct.MonoTextStyle.html
 //! [`ArrayString`]: https://docs.rs/arrayvec/0.4.11/arrayvec/struct.ArrayString.html
 //! [`write!()`]: https://doc.rust-lang.org/nightly/std/macro.write.html
 
-mod font12x16;
-mod font24x32;
-mod font6x12;
-mod font6x8;
-mod font8x16;
+pub mod ascii;
+pub mod latin1;
 mod mono_char_pixels;
 mod mono_text_style;
 
 pub(crate) use mono_char_pixels::MonoCharPixels;
 
-pub use font12x16::Font12x16;
-pub use font24x32::Font24x32;
-pub use font6x12::Font6x12;
-pub use font6x8::Font6x8;
-pub use font8x16::Font8x16;
 pub use mono_text_style::{MonoTextStyle, MonoTextStyleBuilder};
 
 use crate::geometry::Size;